Remove dupes from cachemeifyoucan outputs
`unique` called on a dataframe returns a dataframe with unique rows, just what we need.
Unfortunately, this doesn't really solve the problem, it's merely a band aid.

A better way to solve this problem would be to enforce a UNIQUE KEY constraint in psql layer.
However, it's a little bit tricky and it's not supported by DBI, meaning we'll have to roll our own queries.